split lint tests into multiple files
